Disable source dependencies for plan task
Disable the source dependencies for the plan task as we want to be able
to re-run the plan task even if the file that plan sources do not
change. This is especially true as we work with Workspaces too, so there
may not always be a change between source files and the need to run a
plan.